HW6
