+++ src/test/java/Tests/Darksky/TodayTemp.java
@@ -6,7 +6,7 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 import Pages.Darksky.LandingPage;
 
-public class TemperatureTest {
+public class TodayTemp {
 
     @Test
     public void verifyTodayDate() {
@@ -21,7 +21,6 @@ public class TemperatureTest {
         MyDriver.quitWindows();
     }
 
-
     @Test
     public void verifyTempBars() {
         MyDriver.launchUrlOnNewWindow("https://www.darksky.net/");
@@ -30,7 +29,6 @@ public class TemperatureTest {
         lpage.clickTodayToggle();
         Misc.pause(2);
 
-        // Compare displayed temp vs expanded details temp
         Assert.assertEquals(lpage.getTodayTemp(), lpage.getTodayExtraTemp(), "Test Failed - Displayed temp and expanded details don't match");
         MyDriver.quitWindows();
     }
