+++ src/test/java/Tests/Facebook/Login.java
@@ -1,8 +1,7 @@
 package Tests.Facebook;
 
 import Helper.Misc;
-import Pages.Facebook.LandingPage;
-import Pages.Instagram.InstaLandingPage;
+import Pages.Instagram.LandingPage;
 import Web.MyDriver;
 import org.testng.Assert;
 import org.testng.annotations.Test;
@@ -12,7 +11,7 @@ public class Login {
     @Test
     public void enterLoginInfo() {
         MyDriver.launchUrlOnNewWindow("https://www.facebook.com/");
-        LandingPage lPage = new LandingPage();
+        Pages.Facebook.LandingPage lPage = new Pages.Facebook.LandingPage();
         lPage.enterLoginEmail("rickdeckard@bladerunner.com");
         lPage.enterLoginPassword("nexus6");
         Assert.assertTrue(lPage.isLoginButtonEnabled());
@@ -25,13 +24,13 @@ public class Login {
     @Test
     public void footerLinkVerification() {
         MyDriver.launchUrlOnNewWindow("https://www.facebook.com/");
-        LandingPage lPage = new LandingPage();
+        Pages.Facebook.LandingPage lPage = new Pages.Facebook.LandingPage();
 
         lPage.clickFooterLinks();
         Misc.pause(4);
         lPage.closeTabs();
 
-        InstaLandingPage obj = new InstaLandingPage();
+        LandingPage obj = new LandingPage();
         Assert.assertEquals(obj.instagramPageTitle(), "Instagram", "Test Failed - Title is not Instagram");
         Assert.assertFalse(obj.isInstaLoginBtnEnabled());
         Assert.assertEquals(obj.activeWindowCount(), 1, "Test Failed - more than 1 window handle");
