--- src/test/resources/Facebook/demo.feature
@@ -1,12 +1,0 @@
-Feature: DemoPractice
-
-  Scenario: Testing Cucumber Errors
-    Given I am on amazon.com
-    #When I click on Coupons link
-    Then I verify page title contains coupons
-
-  Scenario: newScenario
-    Given I am on spacex
-    Then I print page title
-
-
